Factor out "PersistTransaction" as a common database utility
